clients/js: refactor worm sui cli commands to prevent type becoming too large
